requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

dogpile.cache
oslo.config>=1.11.0
oslo.log>=1.0.0,<1.1.0
python-slugify
PyYAML>=3.1.0
voluptuous>=0.7
